+++ src/components/display/CardDisplay.tsx
@@ -1,6 +1,7 @@
 import { Card as CardType } from "@/types/interfaces";
 import { SortMode } from "@/types/types";
 import { Link } from "react-router-dom";
+import InfiniteScroll from 'react-infinite-scroll-component';
 import CardFull from "./CardFull";
 import CardImage from "./CardImage";
 import CardList from "./CardList";
@@ -9,35 +10,63 @@ import CardText from "./CardText";
 export default function CardDisplay({
   mode,
   cards,
+  fetchMoreData,
+  hasMore
 }: {
   mode: SortMode;
   cards: CardType[];
+  fetchMoreData: () => void;
+  hasMore: boolean;
 }) {
   switch (mode) {
     default:
     case "image":
       return (
-        <div className="m-4 grid grid-cols-[repeat(auto-fill,minmax(200px,1fr))] justify-center justify-items-center gap-4">
-          {cards.map((card, index) => (
-            <Link to={`/card/${card.attributes.card_code}`} key={index}>
-              <CardImage card={card.attributes} />
-            </Link>
-          ))}
-        </div>
+        <InfiniteScroll
+          dataLength={cards.length}
+          next={fetchMoreData}
+          hasMore={hasMore}
+          loader={<h4>Loading...</h4>}
+          endMessage={
+            <p style={{ textAlign: 'center' }}>
+              <b>Yay! You have seen it all</b>
+            </p>
+          }
+        >
+          <div className="m-4 grid grid-cols-[repeat(auto-fill,minmax(200px,1fr))] justify-center justify-items-center gap-4">
+            {cards.map((card, index) => (
+              <Link to={`/card/${card.attributes.card_code}`} key={index}>
+                <CardImage card={card.attributes} />
+              </Link>
+            ))}
+          </div>
+        </InfiniteScroll>
       );
     case "text":
       return (
-        <div className="m-4 grid grid-cols-[repeat(auto-fill,minmax(350px,1fr))] justify-center justify-items-center gap-2 gap-x-4 gap-y-8">
-          {cards.map((card, index) => (
-            <Link
-              to={`/card/${card.attributes.card_code}`}
-              key={index}
-              className="flex w-full justify-center"
-            >
-              <CardText card={card.attributes} />
-            </Link>
-          ))}
-        </div>
+        <InfiniteScroll
+          dataLength={cards.length}
+          next={fetchMoreData}
+          hasMore={hasMore}
+          loader={<h4>Loading...</h4>}
+          endMessage={
+            <p style={{ textAlign: 'center' }}>
+              <b>Yay! You have seen it all</b>
+            </p>
+          }
+        >
+          <div className="m-4 grid grid-cols-[repeat(auto-fill,minmax(350px,1fr))] justify-center justify-items-center gap-2 gap-x-4 gap-y-8">
+            {cards.map((card, index) => (
+              <Link
+                to={`/card/${card.attributes.card_code}`}
+                key={index}
+                className="flex w-full justify-center"
+              >
+                <CardText card={card.attributes} />
+              </Link>
+            ))}
+          </div>
+        </InfiniteScroll>
       );
     case "list":
       return (
@@ -47,16 +76,28 @@ export default function CardDisplay({
       );
     case "full":
       return (
-        <div>
-          {cards.map((card, index) => (
-            <div
-              key={index}
-              className="relative flex w-full flex-col items-center border-b border-border p-12"
-            >
-              <CardFull card={card.attributes} />
-            </div>
-          ))}
-        </div>
+        <InfiniteScroll
+          dataLength={cards.length}
+          next={fetchMoreData}
+          hasMore={hasMore}
+          loader={<h4>Loading...</h4>}
+          endMessage={
+            <p style={{ textAlign: 'center' }}>
+              <b>Yay! You have seen it all</b>
+            </p>
+          }
+        >
+          <div>
+            {cards.map((card, index) => (
+              <div
+                key={index}
+                className="relative flex w-full flex-col items-center border-b border-border p-12"
+              >
+                <CardFull card={card.attributes} />
+              </div>
+            ))}
+          </div>
+        </InfiniteScroll>
       );
   }
 }
